+++ dsr_agents/CHANGELOG.rst
@@ -21,6 +21,7 @@ X.X.X (XX-XX-2024)
 * Improve formating.
 * Update license to Apache 2.0.
 * Move DSR callback functions to agent_node.
+* Move action_agent to dsr_util.
 
 0.7.0 (19-06-2024)
 ------------------